develop/ai-tools/requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn>=0.28.0
httpx>=0.27.0
pydantic>=2.11.7
pydantic-core>=2.33.2
python-dotenv>=1.1.1
annotated-types>=0.7.0
typing-inspection>=0.4.1
typing-extensions>=4.14.1

# LangGraph & AI
langgraph>=0.2.2
google-generativeai>=0.6.0
google-api-python-client>=2.117.0
grandalf>=0.8

# Web processing & requests
requests>=2.31.0
beautifulsoup4>=4.12.3

# Utilities
tenacity>=8.2.3
cachetools>=5.3.2